LPS-98077 Fixes actions menu for nested fields
