+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/components/FormBuilder/FormBuilder.es.js
@@ -22,7 +22,6 @@ import {Config} from 'metal-state';
 import {pageStructure} from '../../util/config.es';
 import withActionableFields from './withActionableFields.es';
 import withEditablePageHeader from './withEditablePageHeader.es';
-import withHoverableFields from './withHoverableFields.es';
 import withMoveableFields from './withMoveableFields.es';
 import withMultiplePages from './withMultiplePages.es';
 import withResizeableColumns from './withResizeableColumns.es';
@@ -206,7 +205,6 @@ FormBuilderBase.PROPS = {
 export default compose(
 	withActionableFields,
 	withEditablePageHeader,
-	withHoverableFields,
 	withMoveableFields,
 	withMultiplePages,
 	withResizeableColumns
