+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/components/FormBuilder/withActionableFields.es.js
@@ -12,266 +12,17 @@
  * details.
  */
 
-import {ClayActionsDropdown} from 'clay-dropdown';
-import {FormSupport, PagesVisitor} from 'dynamic-data-mapping-form-renderer';
+import {FormSupport} from 'dynamic-data-mapping-form-renderer';
 import dom from 'metal-dom';
 import {EventHandler} from 'metal-events';
 import Component from 'metal-jsx';
-import {Config} from 'metal-state';
 
-import {pageStructure} from '../../util/config.es';
+import FieldActionsDropDown, {
+	getFieldContainer
+} from './FieldActionsDropDown.es';
 import formBuilderProps from './props.es';
 
-const getFieldIndexes = (pages, fieldName) => {
-	const visitor = new PagesVisitor(pages);
-	let indexes = {};
-
-	visitor.mapFields((field, fieldIndex, columnIndex, rowIndex, pageIndex) => {
-		if (field.fieldName === fieldName) {
-			indexes = {columnIndex, pageIndex, rowIndex};
-		}
-	});
-
-	return indexes;
-};
-
-const getFieldRows = ({rows = []}) => {
-	if (typeof rows === 'string') {
-		return JSON.parse(rows);
-	}
-
-	return rows;
-};
-
-const getNestedFieldIndexes = (context, fieldName) => {
-	let indexes = [];
-
-	const visitor = new PagesVisitor(context);
-
-	visitor.mapFields((field, fieldIndex, columnIndex, rowIndex, pageIndex) => {
-		let nestedIndexes = [];
-
-		if (typeof field === 'string') {
-			field = context[0].nestedFields.find(
-				nestedField => nestedField.fieldName === field
-			);
-		}
-
-		if (field && field.fieldName === fieldName) {
-			indexes = [{columnIndex, pageIndex, rowIndex}];
-		}
-		else if (
-			field &&
-			field.fieldName !== fieldName &&
-			field.nestedFields
-		) {
-			nestedIndexes = getNestedFieldIndexes(
-				[
-					{
-						...field,
-						rows: getFieldRows(field)
-					}
-				],
-				fieldName
-			);
-
-			if (nestedIndexes.length) {
-				indexes = [
-					{columnIndex, pageIndex, rowIndex},
-					...nestedIndexes
-				];
-			}
-		}
-	});
-
-	return indexes;
-};
-
-const getFieldContainer = (pages, fieldName) => {
-	const nestedFieldIndexes = getNestedFieldIndexes(pages, fieldName);
-
-	let selector = '';
-
-	nestedFieldIndexes.forEach((fieldIndexes, i) => {
-		const {columnIndex, pageIndex, rowIndex} = fieldIndexes;
-
-		if (i === 0) {
-			selector = [
-				'.ddm-form-page > .row > .col-ddm',
-				`[data-ddm-field-column="${columnIndex}"]`,
-				`[data-ddm-field-page="${pageIndex}"]`,
-				`[data-ddm-field-row="${rowIndex}"]`,
-				'> .ddm-field-container'
-			].join('');
-		}
-		else {
-			selector = [
-				selector,
-				'.ddm-field-container > .ddm-drag > .form-group > .row > .col-ddm',
-				`[data-ddm-field-column="${columnIndex}"]`,
-				`[data-ddm-field-page="${pageIndex}"]`,
-				`[data-ddm-field-row="${rowIndex}"]`,
-				'> .ddm-field-container'
-			].join('');
-		}
-	});
-
-	if (selector) {
-		return document.querySelector(selector);
-	}
-};
-
-class Actions extends Component {
-	created() {
-		this.on('fieldNameChanged', this._handleFieldNameChanged);
-
-		this.expanded = false;
-	}
-
-	render() {
-		const {expanded} = this;
-		const {disabled, items, label, spritemap} = this.props;
-
-		return (
-			<div
-				class="ddm-field-actions-container"
-				onMouseDown={this._handleElementClicked.bind(this)}
-			>
-				<span class="actions-label">{label}</span>
-
-				<ClayActionsDropdown
-					disabled={disabled}
-					events={{
-						expandedChanged: this._handleExpandedChanged.bind(this),
-						itemClicked: this._handleItemClicked.bind(this)
-					}}
-					expanded={expanded}
-					items={items}
-					ref="dropdown"
-					spritemap={spritemap}
-				/>
-			</div>
-		);
-	}
-
-	syncExpanded(expanded) {
-		const {pages} = this.props;
-		const {fieldName} = this.state;
-		const fieldContainer = getFieldContainer(pages, fieldName);
-
-		if (!fieldContainer) {
-			return;
-		}
-
-		if (expanded) {
-			fieldContainer.classList.add('expanded');
-		}
-		else {
-			fieldContainer.classList.remove('expanded');
-		}
-	}
-
-	_handleElementClicked({target}) {
-		const {disabled} = this.props;
-		const {dropdown} = this.refs;
-
-		if (!dropdown.element.contains(target)) {
-			const {dispatch} = this.context;
-			const {fieldName} = this.state;
-			const {pages} = this.props;
-			const indexes = getFieldIndexes(pages, fieldName);
-
-			dispatch('fieldClicked', indexes);
-		}
-		else if (!this.expanded && !disabled) {
-			this.expanded = true;
-		}
-	}
-
-	_handleExpandedChanged({newVal}) {
-		this.expanded = newVal;
-
-		this.syncExpanded(newVal);
-	}
-
-	_handleFieldNameChanged({newVal, prevVal}) {
-		const {pages} = this.props;
-		const {expanded} = this.state;
-		const newFieldContainer = getFieldContainer(pages, newVal);
-		const prevFieldContainer = getFieldContainer(pages, prevVal);
-
-		if (prevFieldContainer && newFieldContainer !== prevFieldContainer) {
-			prevFieldContainer.classList.remove('expanded');
-
-			if (expanded && newFieldContainer) {
-				newFieldContainer.classList.add('expanded');
-			}
-		}
-	}
-
-	_handleItemClicked({
-		data: {
-			item: {action}
-		}
-	}) {
-		const {fieldName} = this.state;
-		const {pages} = this.props;
-		const indexes = getFieldIndexes(pages, fieldName);
-
-		action(indexes);
-
-		this.refs.dropdown.expanded = false;
-	}
-}
-
-Actions.PROPS = {
-	/**
-	 * @default false
-	 * @instance
-	 * @memberof Actions
-	 * @type {!boolean}
-	 */
-
-	disabled: Config.bool().value(false),
-
-	/**
-	 * @default undefined
-	 * @instance
-	 * @memberof Actions
-	 * @type {!string}
-	 */
-
-	label: Config.string(),
-
-	/**
-	 * @default []
-	 * @instance
-	 * @memberof Actions
-	 * @type {?array<object>}
-	 */
-
-	pages: Config.arrayOf(pageStructure).value([]),
-
-	/**
-	 * @default undefined
-	 * @instance
-	 * @memberof Actions
-	 * @type {!string}
-	 */
-
-	spritemap: Config.string().required()
-};
-
-Actions.STATE = {
-	/**
-	 * @default {}
-	 * @instance
-	 * @memberof Actions
-	 * @type {!Object}
-	 */
-
-	fieldName: Config.string()
-};
+const _CSS_HOVERED = 'hovered';
 
 const withActionableFields = ChildComponent => {
 	class ActionableFields extends Component {
@@ -314,7 +65,7 @@ const withActionableFields = ChildComponent => {
 				<div>
 					<ChildComponent {...this.props} />
 
-					<Actions
+					<FieldActionsDropDown
 						disabled={!this.isActionsEnabled()}
 						items={fieldActions}
 						pages={pages}
@@ -324,7 +75,7 @@ const withActionableFields = ChildComponent => {
 						visible={false}
 					/>
 
-					<Actions
+					<FieldActionsDropDown
 						disabled={!this.isActionsEnabled()}
 						items={fieldActions}
 						pages={pages}
@@ -380,41 +131,15 @@ const withActionableFields = ChildComponent => {
 			});
 		}
 
-		_getColumnField(nestedIndexes) {
-			const {pages} = this.props;
-
-			let context = pages;
-			let field;
-
-			nestedIndexes.forEach(indexes => {
-				const {columnIndex, pageIndex, rowIndex} = indexes;
-
-				field = FormSupport.getField(
-					context,
-					field ? 0 : pageIndex,
-					rowIndex,
-					columnIndex
-				);
-
-				context = [
-					{
-						...field,
-						rows: getFieldRows(field)
-					}
-				];
-			});
-
-			return field;
+		_getClosestParent(node) {
+			return dom.closest(node.parentElement, `.ddm-field-container`);
 		}
 
 		_getFieldType(fieldName, field) {
 			const {fieldTypes, pages} = this.props;
 
 			if (!field) {
-				const visitor = new PagesVisitor(pages);
-				field = visitor.findField(
-					fieldItem => fieldItem.fieldName === fieldName
-				);
+				field = FormSupport.findFieldByName(pages, fieldName);
 			}
 
 			return (
@@ -430,41 +155,74 @@ const withActionableFields = ChildComponent => {
 
 			const {delegateTarget} = event;
 
+			this._handleActionsMouseEnter(delegateTarget);
+			this._handleHoverMouseEnter(delegateTarget);
+
+			event.stopPropagation();
+		}
+
+		_handleMouseLeaveField(event) {
+			const {delegateTarget} = event;
+
+			this._handleActionsMouseLeave(delegateTarget);
+			this._handleHoverMouseLeave(delegateTarget);
+
+			event.stopPropagation();
+		}
+
+		_handleActionsMouseEnter(delegateTarget) {
 			if (!delegateTarget.classList.contains('selected')) {
+				const {fieldName} = delegateTarget.dataset;
 				const {hoveredFieldActions} = this.refs;
-				const indexes = FormSupport.getNestedIndexes(
-					dom.closest(delegateTarget, '.col-ddm')
-				);
-				const field = this._getColumnField(indexes);
+				const {pages} = this.props;
+
+				const field = FormSupport.findFieldByName(pages, fieldName);
 
 				if (field) {
-					this.showActions(
-						hoveredFieldActions,
-						field.fieldName,
-						field
-					);
+					this.showActions(hoveredFieldActions, fieldName, field);
 				}
 			}
 		}
 
-		_handleMouseLeaveField(event) {
-			const delegateTarget = dom.closest(
-				event.delegateTarget.parentElement,
-				'.ddm-field-container'
-			);
+		_handleActionsMouseLeave(delegateTarget) {
+			const closestParent = this._getClosestParent(delegateTarget);
 
 			if (
-				delegateTarget &&
-				!delegateTarget.classList.contains('selected')
+				closestParent &&
+				!closestParent.classList.contains('selected')
 			) {
+				const {fieldName} = closestParent.dataset;
 				const {hoveredFieldActions} = this.refs;
-				const indexes = FormSupport.getNestedIndexes(
-					dom.closest(delegateTarget, '.col-ddm')
-				);
-				const field = this._getColumnField(indexes);
+				const {pages} = this.props;
 
-				this.showActions(hoveredFieldActions, field.fieldName, field);
+				const field = FormSupport.findFieldByName(pages, fieldName);
+
+				if (field) {
+					this.showActions(hoveredFieldActions, fieldName, field);
+				}
 			}
+		}
+
+		_handleHoverMouseEnter(delegateTarget) {
+			let closestParent = this._getClosestParent(delegateTarget);
+
+			while (closestParent) {
+				closestParent.classList.remove(_CSS_HOVERED);
+
+				closestParent = this._getClosestParent(closestParent);
+			}
+
+			delegateTarget.classList.add(_CSS_HOVERED);
+		}
+
+		_handleHoverMouseLeave(delegateTarget) {
+			const closestParent = this._getClosestParent(delegateTarget);
+
+			if (closestParent) {
+				closestParent.classList.add(_CSS_HOVERED);
+			}
+
+			delegateTarget.classList.remove(_CSS_HOVERED);
 		}
 	}
